Replace weekly runs with fortnightly ones
Crontab syntax doesn't support schedules based on week numbers, so we
run every week, and use the time since epoch to generate a number that
flips parity every week. If we're on an even week, we set an environment
variable, and only run the other steps if it's set.

The week can be shifted by switching which parity we set the environment
variable on.